fixup! fix(SchemaTree): transfer Show Preview to SchemaTree
