+++ src/containers/Tenant/Schema/SchemaTree/SchemaTree.tsx
@@ -9,6 +9,7 @@ import {useQueryModes} from '../../../../utils/hooks';
 
 import {isChildlessPathType, mapPathTypeToNavigationTreeType} from '../../utils/schema';
 import {getActions} from '../../utils/schemaActions';
+import {getControls} from '../../utils/schemaControls';
 
 interface SchemaTreeProps {
     rootPath: string;
@@ -75,6 +76,10 @@ export function SchemaTree(props: SchemaTreeProps) {
             }}
             fetchPath={fetchPath}
             getActions={getActions(dispatch, {
+                setActivePath: handleActivePathUpdate,
+                setQueryMode,
+            })}
+            renderAdditionalNodeElements={getControls(dispatch, {
                 setActivePath: handleActivePathUpdate,
                 setQueryMode,
             })}
