+++ src/containers/Tenant/utils/SchemaControls.scss
@@ -1,0 +1,5 @@
+.show-preview-button {
+    .yc-button__text {
+        margin: 0 4px;
+    }
+}
